requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.29.1
frozenlist==1.3.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
torch==1.10.2
torchvision==0.11.3
typing-extensions==4.1.0
yarl==1.7.2
